Handle colons in the clues
